Move `Cardano.Fee` to `Cardano.CoinSelection.Fee`.
After applying this change, all public modules are located within a
common namespace: `Cardano.CoinSelection`.
